fix(typedb): share runtime connect parity (#155)
Move direct TypeDB driver access behind a shared runtime crate so the
ORM and server use the same connect-time version gate.

- Share driver construction, HTTP fallback, database lifecycle,
  transactions, and JSON result conversion.
- Rewire ORM and server TypeDB backends to thin adapters over the runtime.
- Add server_version to server config and serverVersion to the Node API.
- Update docs, changelog, and the stale Python package __version__.

Refs #155
Refs #154
