+++ type-bridge-core/crates/node/typescript/index.ts
@@ -497,6 +497,7 @@ export interface NativeRuntime {
     username?: string | null,
     password?: string | null,
     httpPort?: number | null,
+    serverVersion?: string | null,
   ): void;
   connectRustDatabase(
     address: string,
@@ -504,6 +505,7 @@ export interface NativeRuntime {
     username?: string | null,
     password?: string | null,
     httpPort?: number | null,
+    serverVersion?: string | null,
   ): NativeRustDatabase;
 }
 
@@ -521,6 +523,8 @@ export interface RustDatabaseConnectOptions {
   password?: string | null;
   /** Port of the TypeDB HTTP API used for the connect-time version probe. */
   httpPort?: number;
+  /** Exact TypeDB server version; skips HTTP probing for gRPC-only deployments. */
+  serverVersion?: string | null;
 }
 
 export interface EnsureDatabaseOptions {
@@ -528,6 +532,8 @@ export interface EnsureDatabaseOptions {
   password?: string | null;
   /** Port of the TypeDB HTTP API used for the connect-time version probe. */
   httpPort?: number;
+  /** Exact TypeDB server version; skips HTTP probing for gRPC-only deployments. */
+  serverVersion?: string | null;
 }
 
 /**
@@ -548,6 +554,7 @@ export function ensureDatabase(
     options?.username ?? null,
     options?.password ?? null,
     options?.httpPort ?? null,
+    options?.serverVersion ?? null,
   );
 }
 
@@ -751,6 +758,7 @@ export class RustDatabase {
         parsed.options.username ?? null,
         parsed.options.password ?? null,
         parsed.options.httpPort ?? null,
+        parsed.options.serverVersion ?? null,
       ),
     );
   }
