+++ type-bridge-core/crates/orm/Cargo.toml
@@ -17,25 +17,21 @@ path = "src/bin/codegen.rs"
 
 [features]
 default = ["band7", "band8"]
-typedb = ["dep:futures"]
-band7 = ["typedb", "dep:type-bridge-typedb-driver-b7"]
-band8 = ["typedb", "dep:typedb-driver"]
+typedb = ["dep:type-bridge-typedb-runtime"]
+band7 = ["typedb", "type-bridge-typedb-runtime/band7"]
+band8 = ["typedb", "type-bridge-typedb-runtime/band8"]
 derive = ["dep:type-bridge-orm-derive"]
 integration-tests = ["derive", "band7", "band8"]
 
 [dependencies]
 type-bridge-core-lib = { path = "../core", version = "1.5.2-rc1" }
+type-bridge-typedb-runtime = { path = "../typedb-runtime", version = "1.5.2-rc1", default-features = false, optional = true }
 tokio = { version = "1", features = ["full"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 thiserror = "2"
 tracing = "0.1"
 type-bridge-orm-derive = { path = "../orm-derive", version = "1.5.2-rc1", optional = true }
-
-# Optional: real TypeDB backend
-typedb-driver = { version = "3", optional = true }
-type-bridge-typedb-driver-b7 = { path = "../../vendor/typedb-driver-b7", version = "=3.8.1", optional = true }
-futures = { version = "0.3", optional = true }
 
 [dev-dependencies]
 tokio = { version = "1", features = ["full", "test-util"] }
