+++ type-bridge-core/crates/server/Cargo.toml
@@ -18,14 +18,15 @@ required-features = ["typedb", "axum-transport"]
 
 [features]
 default = ["band7", "band8", "axum-transport"]
-typedb = ["dep:futures"]
-band7 = ["typedb", "dep:type-bridge-typedb-driver-b7"]
-band8 = ["typedb", "dep:typedb-driver"]
+typedb = ["dep:type-bridge-typedb-runtime"]
+band7 = ["typedb", "type-bridge-typedb-runtime/band7"]
+band8 = ["typedb", "type-bridge-typedb-runtime/band8"]
 axum-transport = ["dep:axum", "dep:tower-http"]
 test-helpers = []
 
 [dependencies]
 type-bridge-core-lib = { path = "../core", version = "1.5.2-rc1" }
+type-bridge-typedb-runtime = { path = "../typedb-runtime", version = "1.5.2-rc1", default-features = false, optional = true }
 tokio = { version = "1", features = ["full"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
@@ -36,10 +37,6 @@ uuid = { version = "1", features = ["v4"] }
 chrono = { version = "0.4", features = ["serde"] }
 thiserror = "2"
 clap = { version = "4", features = ["derive"] }
-# Optional: TypeDB backend
-typedb-driver = { version = "3", optional = true }
-type-bridge-typedb-driver-b7 = { path = "../../vendor/typedb-driver-b7", version = "=3.8.1", optional = true }
-futures = { version = "0.3", optional = true }
 
 # Optional: Axum HTTP transport
 axum = { version = "0.8", optional = true }
