+++ type-bridge-core/crates/server/src/typedb/client.rs
@@ -132,141 +132,6 @@ pub(crate) fn parse_transaction_type(tx_type: &str) -> Result<TransactionType, P
     }
 }
 
-/// Convert a band-7 TypeDB concept to a JSON value.
-///
-/// Output shape is identical to [`concept_to_json_b8`] for all common concepts.
-#[cfg(feature = "band7")]
-pub(crate) fn concept_to_json_b7(
-    concept: &type_bridge_typedb_driver_b7::concept::Concept,
-) -> serde_json::Value {
-    let mut obj = serde_json::Map::new();
-    obj.insert(
-        "category".to_string(),
-        serde_json::Value::String(concept.get_category().name().to_string()),
-    );
-    obj.insert(
-        "label".to_string(),
-        serde_json::Value::String(concept.get_label().to_string()),
-    );
-    if let Some(iid) = concept.try_get_iid() {
-        obj.insert(
-            "iid".to_string(),
-            serde_json::Value::String(iid.to_string()),
-        );
-    }
-    if let Some(value) = concept.try_get_value() {
-        obj.insert("value".to_string(), value_to_json_b7(value));
-    }
-    if let Some(value_type) = concept.try_get_value_type() {
-        obj.insert(
-            "value_type".to_string(),
-            serde_json::Value::String(value_type.name().to_string()),
-        );
-    }
-    serde_json::Value::Object(obj)
-}
-
-/// Convert a band-7 TypeDB value to a JSON value.
-#[cfg(feature = "band7")]
-#[cfg_attr(coverage_nightly, coverage(off))]
-fn value_to_json_b7(value: &type_bridge_typedb_driver_b7::concept::Value) -> serde_json::Value {
-    if let Some(b) = value.get_boolean() {
-        return serde_json::Value::Bool(b);
-    }
-    if let Some(i) = value.get_integer() {
-        return serde_json::json!(i);
-    }
-    if let Some(d) = value.get_double() {
-        return serde_json::json!(d);
-    }
-    if let Some(s) = value.get_string() {
-        return serde_json::Value::String(s.to_string());
-    }
-    if let Some(date) = value.get_date() {
-        return serde_json::Value::String(date.to_string());
-    }
-    if let Some(dt) = value.get_datetime() {
-        return serde_json::Value::String(dt.to_string());
-    }
-    if let Some(dt_tz) = value.get_datetime_tz() {
-        return serde_json::Value::String(dt_tz.to_string());
-    }
-    if let Some(dec) = value.get_decimal() {
-        return serde_json::Value::String(dec.to_string());
-    }
-    if let Some(dur) = value.get_duration() {
-        return serde_json::Value::String(dur.to_string());
-    }
-    serde_json::Value::String(value.to_string())
-}
-
-/// Convert a band-8 TypeDB concept to a JSON value.
-///
-/// Output shape is identical to [`concept_to_json_b7`] for all common concepts.
-#[cfg(feature = "band8")]
-pub(crate) fn concept_to_json_b8(concept: &typedb_driver::concept::Concept) -> serde_json::Value {
-    let mut obj = serde_json::Map::new();
-    obj.insert(
-        "category".to_string(),
-        serde_json::Value::String(concept.get_category().name().to_string()),
-    );
-    obj.insert(
-        "label".to_string(),
-        serde_json::Value::String(concept.get_label().to_string()),
-    );
-    if let Some(iid) = concept.try_get_iid() {
-        obj.insert(
-            "iid".to_string(),
-            serde_json::Value::String(iid.to_string()),
-        );
-    }
-    if let Some(value) = concept.try_get_value() {
-        obj.insert("value".to_string(), value_to_json_b8(value));
-    }
-    if let Some(value_type) = concept.try_get_value_type() {
-        obj.insert(
-            "value_type".to_string(),
-            serde_json::Value::String(value_type.name().to_string()),
-        );
-    }
-    serde_json::Value::Object(obj)
-}
-
-/// Convert a band-8 TypeDB value to a JSON value.
-#[cfg(feature = "band8")]
-#[cfg_attr(coverage_nightly, coverage(off))]
-pub(crate) fn value_to_json_b8(value: &typedb_driver::concept::Value) -> serde_json::Value {
-    if let Some(b) = value.get_boolean() {
-        return serde_json::Value::Bool(b);
-    }
-    if let Some(i) = value.get_integer() {
-        return serde_json::json!(i);
-    }
-    if let Some(d) = value.get_double() {
-        return serde_json::json!(d);
-    }
-    if let Some(s) = value.get_string() {
-        return serde_json::Value::String(s.to_string());
-    }
-    if let Some(date) = value.get_date() {
-        return serde_json::Value::String(date.to_string());
-    }
-    if let Some(dt) = value.get_datetime() {
-        return serde_json::Value::String(dt.to_string());
-    }
-    if let Some(dt_tz) = value.get_datetime_tz() {
-        return serde_json::Value::String(dt_tz.to_string());
-    }
-    if let Some(dec) = value.get_decimal() {
-        return serde_json::Value::String(dec.to_string());
-    }
-    if let Some(dur) = value.get_duration() {
-        return serde_json::Value::String(dur.to_string());
-    }
-    // Fallback: use Display representation
-    serde_json::Value::String(value.to_string())
-}
-
 #[cfg(test)]
 #[cfg(feature = "band8")]
 #[cfg_attr(coverage_nightly, coverage(off))]
@@ -276,12 +141,6 @@ mod tests {
     use std::pin::Pin;
     use std::sync::Arc;
     use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-
-    use typedb_driver::IID;
-    use typedb_driver::concept::value::{Decimal, Duration, TimeZone};
-    use typedb_driver::concept::{
-        Attribute, AttributeType, Concept, Entity, EntityType, Value, ValueType,
-    };
 
     use super::*;
     use crate::error::PipelineError;
@@ -951,204 +810,6 @@ mod tests {
     }
 
     // =============================================
-    // value_to_json tests
-    // =============================================
-
-    #[test]
-    fn value_to_json_boolean_true() {
-        let value = Value::Boolean(true);
-        let json = value_to_json_b8(&value);
-        assert_eq!(json, serde_json::Value::Bool(true));
-    }
-
-    #[test]
-    fn value_to_json_boolean_false() {
-        let value = Value::Boolean(false);
-        let json = value_to_json_b8(&value);
-        assert_eq!(json, serde_json::Value::Bool(false));
-    }
-
-    #[test]
-    fn value_to_json_integer() {
-        let value = Value::Integer(42);
-        let json = value_to_json_b8(&value);
-        assert_eq!(json, serde_json::json!(42));
-    }
-
-    #[test]
-    fn value_to_json_integer_negative() {
-        let value = Value::Integer(-100);
-        let json = value_to_json_b8(&value);
-        assert_eq!(json, serde_json::json!(-100));
-    }
-
-    #[test]
-    fn value_to_json_double() {
-        let value = Value::Double(3.15);
-        let json = value_to_json_b8(&value);
-        assert_eq!(json, serde_json::json!(3.15));
-    }
-
-    #[test]
-    fn value_to_json_string() {
-        let value = Value::String("hello".to_string());
-        let json = value_to_json_b8(&value);
-        assert_eq!(json, serde_json::Value::String("hello".to_string()));
-    }
-
-    #[test]
-    fn value_to_json_string_empty() {
-        let value = Value::String(String::new());
-        let json = value_to_json_b8(&value);
-        assert_eq!(json, serde_json::Value::String(String::new()));
-    }
-
-    #[test]
-    fn value_to_json_date() {
-        let date = chrono::NaiveDate::from_ymd_opt(2024, 1, 15).unwrap();
-        let value = Value::Date(date);
-        let json = value_to_json_b8(&value);
-        assert_eq!(json, serde_json::Value::String("2024-01-15".to_string()));
-    }
-
-    #[test]
-    fn value_to_json_datetime() {
-        let dt = chrono::NaiveDate::from_ymd_opt(2024, 1, 15)
-            .unwrap()
-            .and_hms_opt(10, 30, 0)
-            .unwrap();
-        let value = Value::Datetime(dt);
-        let json = value_to_json_b8(&value);
-        let s = json.as_str().unwrap();
-        assert!(s.contains("2024-01-15"));
-    }
-
-    #[test]
-    fn value_to_json_decimal() {
-        let dec = Decimal::new(42, 0);
-        let value = Value::Decimal(dec);
-        let json = value_to_json_b8(&value);
-        assert!(json.is_string());
-    }
-
-    #[test]
-    fn value_to_json_duration() {
-        let dur = Duration::new(1, 2, 3_000_000_000);
-        let value = Value::Duration(dur);
-        let json = value_to_json_b8(&value);
-        assert!(json.is_string());
-    }
-
-    #[test]
-    fn value_to_json_datetime_tz() {
-        use chrono::TimeZone as _;
-        let tz = TimeZone::Fixed(chrono::FixedOffset::east_opt(3600).unwrap());
-        let dt = tz.with_ymd_and_hms(2024, 6, 15, 12, 30, 0).unwrap();
-        let value = Value::DatetimeTZ(dt);
-        let json = value_to_json_b8(&value);
-        let s = json.as_str().unwrap();
-        assert!(s.contains("2024"));
-    }
-
-    // =============================================
-    // concept_to_json tests
-    // =============================================
-
-    #[test]
-    fn concept_to_json_entity_type() {
-        let concept = Concept::EntityType(EntityType {
-            label: "person".to_string(),
-        });
-        let json = concept_to_json_b8(&concept);
-        assert_eq!(json["category"], "EntityType");
-        assert_eq!(json["label"], "person");
-        assert!(json.get("iid").is_none());
-        assert!(json.get("value").is_none());
-    }
-
-    #[test]
-    fn concept_to_json_attribute_type() {
-        let concept = Concept::AttributeType(AttributeType {
-            label: "name".to_string(),
-            value_type: Some(ValueType::String),
-        });
-        let json = concept_to_json_b8(&concept);
-        assert_eq!(json["category"], "AttributeType");
-        assert_eq!(json["label"], "name");
-        assert_eq!(json["value_type"], "string");
-    }
-
-    #[test]
-    fn concept_to_json_value_boolean() {
-        let concept = Concept::Value(Value::Boolean(true));
-        let json = concept_to_json_b8(&concept);
-        assert_eq!(json["category"], "Value");
-        assert_eq!(json["value"], true);
-    }
-
-    #[test]
-    fn concept_to_json_value_integer() {
-        let concept = Concept::Value(Value::Integer(42));
-        let json = concept_to_json_b8(&concept);
-        assert_eq!(json["value"], 42);
-    }
-
-    #[test]
-    fn concept_to_json_value_string() {
-        let concept = Concept::Value(Value::String("hello".to_string()));
-        let json = concept_to_json_b8(&concept);
-        assert_eq!(json["value"], "hello");
-    }
-
-    #[test]
-    fn concept_to_json_entity_with_iid() {
-        let iid: IID = vec![0x01, 0x02, 0x03].into();
-        let concept = Concept::Entity(Entity {
-            iid,
-            type_: Some(EntityType {
-                label: "person".to_string(),
-            }),
-        });
-        let json = concept_to_json_b8(&concept);
-        assert_eq!(json["category"], "Entity");
-        assert_eq!(json["label"], "person");
-        // IID should be present
-        let iid_str = json["iid"].as_str().unwrap();
-        assert!(iid_str.starts_with("0x"));
-    }
-
-    #[test]
-    fn concept_to_json_attribute_with_value() {
-        let iid: IID = vec![0xAA, 0xBB].into();
-        let concept = Concept::Attribute(Attribute {
-            iid,
-            value: Value::String("hello".to_string()),
-            type_: Some(AttributeType {
-                label: "name".to_string(),
-                value_type: Some(ValueType::String),
-            }),
-        });
-        let json = concept_to_json_b8(&concept);
-        assert_eq!(json["category"], "Attribute");
-        assert_eq!(json["label"], "name");
-        // Attribute IID is not exposed via try_get_iid()
-        assert!(json.get("iid").is_none());
-        assert_eq!(json["value"], "hello");
-        assert_eq!(json["value_type"], "string");
-    }
-
-    #[test]
-    fn concept_to_json_attribute_type_without_value_type() {
-        let concept = Concept::AttributeType(AttributeType {
-            label: "abstract_attr".to_string(),
-            value_type: None,
-        });
-        let json = concept_to_json_b8(&concept);
-        assert_eq!(json["label"], "abstract_attr");
-        assert!(json.get("value_type").is_none());
-    }
-
-    // =============================================
     // QueryExecutor trait impl tests
     // =============================================
 
@@ -1195,6 +856,7 @@ mod tests {
             username: "admin".to_string(),
             password: "password".to_string(),
             http_port: 8000,
+            server_version: None,
         };
         let result = TypeDBClient::connect(&config).await;
         assert!(result.is_err());
@@ -1213,6 +875,7 @@ mod tests {
                 .ok()
                 .and_then(|port| port.parse().ok())
                 .unwrap_or(8000),
+            server_version: std::env::var("TYPEDB_SERVER_VERSION").ok(),
         }
     }
